"Refactor code for performance and maintainability"
Completely refactored FileCrudRepository, App, User, and CrudRepository classes to improve performance and readability. CRUD operations in the FileCrudRepository now use Java streams for improved performance and maintainability. Added UserMapper class to handle user Persistence and data formatting. Renamed FileCrudRepository to FileUserCrudRepository to be more representative of its function. Reworked the App class to simplify it and reflect the changes within the other classes. Replaced the serialization demo code with appropriate calls to the FileUserCrudRepository class. User class was updated to include a salary field and improved equals and hashCode methods. Finally, changed all id's from long to int in CrudRepository to correct data inconsistencies.

+++ src/main/java/io/dz/CrudRepository.java
@@ -6,13 +6,13 @@ import java.util.Optional;
 
 public interface CrudRepository<T> {
 
-    Optional<T> findById(long id) throws IOException;
+    Optional<T> findById(int id) throws IOException;
 
     Optional<T> findByObject(T t) throws IOException;
 
     void save(T t) throws IOException;
 
-    void delete(long id) throws IOException;
+    void delete(int id) throws IOException;
 
     List<T> findAll() throws IOException;
 }
